service-otp-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0
pyotp==2.9.0